# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: AgentWire API
  version: 2.0.0
  description: >
    API for AgentWire: AI traffic ingestion, security scanning, replay,
    RBAC/SSO-authenticated access, and audit logging. Single organization per
    self-hosted instance (see docs/architecture.md).
security:
  - BearerAuth: []
paths:
...
  /v1/traces:
    post:
      summary: Ingest a new AI packet or trace. Runs the security scanner inline.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIPacket'
      responses:
        '202':
          description: Packet accepted, findings recorded synchronously.
        '401': { description: Missing, invalid, or revoked API key. }
...
  /v1/packets/{id}:
    get:
      summary: Fetch a single packet by id (org-scoped - 404 if it belongs to another org).
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AIPacket' }
        '404': { description: Not found (including packets that belong to a different organization). }
  /v1/packets/{id}/findings:
    get:
      summary: Security findings for a single packet.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SecurityFinding' }
        '404': { description: Packet not found. }
  /v1/packets/{id}/replay:
    post:
      summary: Re-send a stored prompt to a configured OpenAI-compatible provider.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                model: { type: string }
                temperature: { type: number }
      responses:
        '200':
          description: Replay succeeded; a new linked packet was created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  replayId: { type: string, format: uuid }
                  newPacketId: { type: string, format: uuid }
                  response: { type: string }
                  promptTokens: { type: integer }
                  completionTokens: { type: integer }
                  latencyMs: { type: number }
                  cost: { type: number }
        '404': { description: Packet not found. }
        '422': { description: "No Replay:BaseUrl configured - not a fake success." }
        '502': { description: Configured provider errored or timed out. }
  /v1/packets/{id}/replays:
    get:
      summary: Replay attempt history for a packet (including failures).
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ReplayResult' }
  /v1/security/findings:
    get:
      summary: Org-wide security findings, paginated.
      parameters:
        - { name: type, in: query, schema: { type: string } }
        - { name: severity, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 50 } }
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SecurityFinding' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Self-issued JWT (local login, or find-or-provisioned via OIDC/SAML).
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Per-organization ingestion key, used only by POST /v1/traces.
  schemas:
    AIPacket:
      type: object
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        replayOfPacketId: { type: string, format: uuid, nullable: true }
        traceId: { type: string }
        agentId: { type: string }
        modelProvider: { type: string }
        modelName: { type: string }
        systemPrompt: { type: string }
        userPrompt: { type: string }
        llmResponse: { type: string }
        promptTokens: { type: integer }
        completionTokens: { type: integer }
        cost: { type: number }
        latencyMs: { type: number }
        createdAt: { type: string, format: date-time }
      required:
        - traceId
        - modelName
    SecurityFinding:
      type: object
      properties:
        id: { type: string, format: uuid }
        aiPacketId: { type: string, format: uuid }
        findingType: { type: string, enum: [PromptInjection, PiiEmail, PiiPhone, PiiCreditCard, PiiSsn] }
        severity: { type: string, enum: [Low, Medium, High] }
        location: { type: string, enum: [SystemPrompt, UserPrompt, LlmResponse] }
        ruleId: { type: string }
        matchedTextMasked: { type: string, description: "Masked - never the raw matched text." }
        createdAt: { type: string, format: date-time }
    ReplayResult:
      type: object
      properties:
        id: { type: string, format: uuid }
        originalPacketId: { type: string, format: uuid }
        newPacketId: { type: string, format: uuid, nullable: true }
        status: { type: string, enum: [Pending, Succeeded, Failed] }
        errorMessage: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...